fix(android): WaveformAccumulator Int-Overflow crasht Trim-Editor bei langen Dateien
Beim Öffnen des Edit-Sheets/Trim-Editors einer Meditation > ~22 s crashte die App
(ArrayIndexOutOfBoundsException, negativer Bucket-Index). Ursache: die Bucket-Zuordnung
`frame * bucketCount` rechnete in 32-bit Int und lief bei langen Dateien (z.B. 32 min ≈
84 Mio. Frames × 2200 Buckets) über → negativer Index. iOS war nicht betroffen (Swift Int
ist 64-bit); der Phase-C-Spike rechnete bereits in Long, die Unit-Tests nutzten kleine
Frame-Zahlen — daher unentdeckt.

Fix: Bucket-Arithmetik in Long. Zusätzlich kapselt der MediaCodec-Decode-Loop jetzt
IllegalStateException/CodecException als WaveformGenerationException, damit echte
Decode-Fehler wie dokumentiert auf die Fallback-Linie führen statt zu crashen.

Test: WaveformAccumulator mit 1,1 Mio. Frames (über der Overflow-Schwelle). On-Device
verifiziert (Pixel 8 Emulator, 32-min-Datei): Edit-Sheet + Trim-Editor öffnen sauber.
1203 Unit-Tests grün, make check grün.

+++ android/app/src/main/kotlin/com/stillmoment/domain/models/WaveformAccumulator.kt
@@ -59,9 +59,14 @@ class WaveformAccumulator(
         return MeditationWaveform(peaks.map { it / globalMax })
     }
 
-    /** Maps a global frame index to its bucket, distributing frames evenly. */
+    /**
+     * Maps a global frame index to its bucket, distributing frames evenly.
+     *
+     * Uses 64-bit arithmetic: a long file has tens of millions of frames, so
+     * `frame * bucketCount` overflows a 32-bit Int (negative bucket index → crash).
+     */
     private fun bucketIndex(frame: Int): Int {
-        val rawIndex = frame * bucketCount / totalFrameCount
+        val rawIndex = (frame.toLong() * bucketCount / totalFrameCount).toInt()
         return minOf(rawIndex, bucketCount - 1)
     }
 }
